modified the bill calculator to take decimal prices as well

+++ Mini Projects/Bill Calculator.py
@@ -1,7 +1,7 @@
 # inputs
 print("Please enter the following information:")
 name=input("Item name: ")
-price=int(input(f"Unit price of {name}: "))
+price=float(input(f"Unit price of {name}: "))
 qty=int(input("Quantity: "))
 
 # operations
